Fixed Line Graphs (displays w prop < LOD)
I have fixed the formatting issue that was preventing the display of line graphs.  However, I am allowing the line graph to plot points for geomeans even when the proportion of results below the LOD are >40%.  This may need another fix if it is required we remove these points from visualization.